+++ Assets/VRM10/Runtime/IO/VrmAnimationImporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0381f410b38d5184baa6f28213608810
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
