+++ Assets/VRM10_Samples/VRM10Viewer/Motions/BvhMotion.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8bed75b815cb24843a50063a9dd7376e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
