+++ Assets/VRM10_Samples/VRM10Viewer/Motions/IMotion.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bd8db8a4bdc3e7f40a0a7a1bcd554137
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
