+++ Assets/VRM10_Samples/VRM10Viewer/Motions/VrmAnimation.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 08f06b12436bf594b9982c6a8ec71374
+guid: f0180f844d447bc45889d6c72a2d4ec4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
